App compatibility and testing.
The app appears to find and pluck our visualisations and run them inside an iframe.

A script is added which creates a "* {}" style, which makes all the elements have the same font and font-size.

This is terrible and we must destroy it. Prelaunch does that now. However, in order to destroy this, prelaunch must run *after* the script, which means it must be deferred, which means everything else must also be deferred.